# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShopFlow Product Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  imageUrl: string;
  images: string[];
  category: string;
  categoryId: number;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  stockCount: number;
  isFeatured: boolean;
  isNew: boolean;
  tags: string[];
  createdAt: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  productCount: number;
  imageUrl: string;
}

export interface StoreStats {
  totalProducts: number;
  totalCategories: number;
  featuredCount: number;
  avgRating: number;
}

export interface CartItem {
  productId: number;
  product: Product;
  quantity: number;
}

export interface Cart {
  items: CartItem[];
  totalItems: number;
  subtotal: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  sort?: ListProductsSort;
  minPrice?: number;
  maxPrice?: number;
};

export type ListProductsSort =
  (typeof ListProductsSort)[keyof typeof ListProductsSort];

export const ListProductsSort = {
  price_asc: "price_asc",
  price_desc: "price_desc",
  newest: "newest",
  popular: "popular",
} as const;