# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShopFlow Product Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  imageUrl: string;
  images: string[];
  category: string;
  categoryId: number;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  stockCount: number;
  isFeatured: boolean;
  isNew: boolean;
  tags: string[];
  createdAt: string;
}